+++ NEWS.rst
@@ -1,3 +1,15 @@
+Changes in 3.2.4 (released ??/??/2008)
+--------------------------------------
+
+*	The following functions have been added to :mod:`ll.xist.css`:
+	:func:`parsestring`, :func:`parsestream`, :func:`parsefile`, :func:`parseurl`
+	and :func:`write`. They parse CSS resources and are similar to the XML/HTML
+	parsing function in that they apply the specified base URL to all URL in the
+	style sheet.
+
+*	:mod:`cssutils` 0.9.5b1 is required now.
+
+
 Changes in 3.2.3 (released 03/04/2008)
 --------------------------------------